Add SQL module with dependencies and initial classes
This commit introduces a new `sql` module to the existing project. This module includes a Maven pom.xml configuration with necessary dependencies on SQL drivers, such as MariaDB, MySQL, PostgreSQL, as well as exposed dependencies. It also includes several classes (`SqlConfig`, `SqlDriver`, `AbstractRepository`, `SqlConnectionFactory`) essential for setting up a SQL connection and interaction with a database. This was needed to extend the functionality of the existing app, allowing it to interact with SQL databases.

Also, the root pom.xml file was updated, adding the version property for the exposed dependencies and mentioning the newly added sql module. The .idea/vcs.xml file was also added to take care of the new modules.